requirements.txt
anyio==4.6.0
APScheduler==3.10.4
certifi==2024.8.30
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.1
peewee==3.17.7
python-dotenv==1.0.1
python-telegram-bot==21.6
pytz==2024.2
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
tzlocal==5.2
